bugfixes and code rearrangements
